Add templatetags with new filter
templatetags folder now in food_track. It is configured in setttings.py as TEMPLATES[0]['OPTIONS']['libraries']['templatetags']. It is loaded in search_results.html, and the filter process_qry processes query to be returned in the search box to show the previous search query. Works with different page numbers and a few special characters.